Patch strlen

+++ src/util2.c
@@ -6,7 +6,7 @@
 /*   By: smamalig <smamalig@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/08 23:29:07 by smamalig          #+#    #+#             */
-/*   Updated: 2025/05/08 23:37:22 by smamalig         ###   ########.fr       */
+/*   Updated: 2025/05/09 21:47:35 by smamalig         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ int	__ft_printf_isdigit(char c)
 
 size_t	__ft_printf_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
 	while (s[i++ + 1])
